# Conflicts:
#	Template_Project/Src/Controller.cpp
#	Template_Project/Src/Game.cpp

+++ Template_Project/Src/Cells.cpp
@@ -14,7 +14,7 @@ void Cells::initializeCells()
 												{245,375,0},
 												{375,375,0}
 										};
-	for(int i = 0; i<9; i++)
+	for(BYTE i = 0; i<9; i++)
 	{
 				cells[i] = newCells[i];
 	}
@@ -22,15 +22,15 @@ void Cells::initializeCells()
 
 void Cells::initializeDefaultRows()
 {
-	topRow = new short[3];
-	centerRow = new short[3];
-	bottomRow = new short[3];
-	leftColumn = new short[3];
-	centerColumn = new short[3];
-	rightColumn = new short[3];
-	downDiagonal = new short[3];
-	upDiagonal = new short[3];
-	for(short i = 0; i<3; i++)
+	topRow = new BYTE[3];
+	centerRow = new BYTE[3];
+	bottomRow = new BYTE[3];
+	leftColumn = new BYTE[3];
+	centerColumn = new BYTE[3];
+	rightColumn = new BYTE[3];
+	downDiagonal = new BYTE[3];
+	upDiagonal = new BYTE[3];
+	for(BYTE i = 0; i<3; i++)
 	{
 		topRow[i] = i;
 		centerRow[i] = 3+i;
@@ -52,8 +52,8 @@ Cells::Cells()
 
 bool Cells::rowIsComplete()
 {
-	short rowValues[8] = {0,0,0,0,0,0,0,0};
-	for(int i = 0;i<3;i++)
+	BYTE rowValues[8] = {0,0,0,0,0,0,0,0};
+	for(BYTE i = 0;i<3;i++)
 	{
 		rowValues[0] += cells[topRow[i]].player == 0 ? 0 : cells[topRow[i]].player+2;
 		rowValues[1] += cells[centerRow[i]].player == 0 ? 0 : cells[centerRow[i]].player+2;
